requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
prophet==1.1.5
pydantic==2.6.4
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
httpx==0.27.0
apscheduler==3.10.4